[CHG] ---------------------------------------- Changed routes so they are version-controlled
